requirements.txt
analytics-python==1.2.9
appnope==0.1.2
arch==4.18
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.3
bleach==3.3.0
Brotli==1.0.9
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cvxpy==1.1.10
cycler==0.10.0
Cython==0.29.22
dash==1.19.0
dash-bootstrap-components==0.11.3
dash-core-components==1.15.0
dash-cytoscape==0.2.0
dash-dangerously-set-inner-html==0.0.2
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.7.1
demos==0.0.1
ecos==2.0.7.post1
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
getmac==0.8.2
greenlet==1.1.1
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
importlib-metadata==3.7.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==5.6.1
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
notebook==6.2.0
numpy==1.20.1
osqp==0.6.2.post0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.2
parso==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
property-cached==1.6.4
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qdldl==0.1.5.post0
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
retrying==1.3.3
Riskfolio-Lib==0.1.4
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
scs==2.1.2
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
statsmodels==0.12.2
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.17
torch==1.9.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.10.0.2
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
XlsxWriter==1.3.7
zipp==3.4.1
